Simplify the division between manager and processor
