+++ src/Interfaces/PaymentGateway.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace rkujawa\LaravelPaymentGateway\Interfaces;
+
+interface PaymentGateway extends PaymentManagement, PaymentProcessing
+{
+    
+}
