--- src/PaymentProcessor.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace rkujawa\LaravelPaymentGateway;
-
-use rkujawa\LaravelPaymentGateway\Interfaces\PaymentProcessing;
-use rkujawa\LaravelPaymentGateway\Traits\ProcessesPayments;
-
-class PaymentProcessor extends PaymentService implements PaymentProcessing
-{
-    use ProcessesPayments;
-
-    protected $service = PaymentService::PROCESSOR;
-}
